Split the project into modules:
- The core module, providing the basic building
blocks
- The JUnit 4 Runner module, providing JUnit Vintage
support
