+++ core/src/main/java/javadoctest/internal/ExtractedDocTest.java
