+++ junit-vintage-runner/src/test/java/javadoctest/vintage/VariablesDocTest.java
@@ -1,5 +1,6 @@
-package javadoctest;
+package javadoctest.vintage;
 
+import javadoctest.DocSnippet;
 import org.junit.runner.RunWith;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -9,7 +10,7 @@ import static org.hamcrest.Matchers.equalTo;
 public class VariablesDocTest
 {
     /**
-     * <pre class="doctest:javadoctest.VariablesDocTest#testCanReadVar">
+     * <pre class="doctest:javadoctest.vintage.VariablesDocTest#testCanReadVar">
      * String aString = "Hello, world!";
      * </pre>
      */
